shared/api,lxd/operations,lxc/operation: Use int64 for ChildCount to match OpenAPI schema

+++ lxd/operations/operations.go
@@ -778,7 +778,7 @@ func (op *Operation) Render() (string, *api.Operation) {
 		Location:    op.location,
 		Err:         op.err,
 		ErrCode:     op.errCode,
-		ChildCount:  len(op.children),
+		ChildCount:  int64(len(op.children)),
 	}
 
 	requestor := op.Requestor()
